Moving astrometry index downloads to the script for that purpose. Cleaning up the `panoptes.utils.data` namespace.
