--- panoptes/utils/data/__init__.py
@@ -1,1 +1,0 @@
-from .assets import Downloader
